requirements.txt
astroid==3.3.10
coverage==7.8.2
dill==0.4.0
exceptiongroup==1.3.0
iniconfig==2.1.0
isort==6.0.1
mccabe==0.7.0
numpy==2.0.2
packaging==25.0
pandas==2.3.0
platformdirs==4.3.8
pluggy==1.6.0
Pygments==2.19.1
pylint==3.3.7
pytest==8.4.0
pytest-cov==6.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
tomli==2.2.1
tomlkit==0.13.3
typing_extensions==4.14.0
tzdata==2025.2
faker